Update snapchat_agent.php
separate isMedia into isImage and isVideo functions which can be used independently to identify if the media is an image or a video. The isMedia function relies on the individual functions and allows a consumer of this class to use these headers which are much more accurate than the media_type attribute which I have (anecdotally) found to be unreliable.